feat: creating storage file (#21)
* creating storage file

* adding property test for the storage and updating the wasmer tests

* updating the change log

* adding header to the storage file

* loading storage file and read code from the file

+++ tanour/src/page.rs
@@ -1,12 +1,12 @@
 #[derive(Debug)]
 pub struct Page {
-    pub offset: usize,
-    pub length: usize,
+    pub offset: u32,
+    pub length: u32,
     pub data: Vec<u8>,
     pub updated: bool,
 }
 impl Page {
-    pub fn new(offset: usize, length: usize, data: Vec<u8>) -> Self {
+    pub fn new(offset: u32, length: u32, data: Vec<u8>) -> Self {
         Page {
             offset,
             length,
